requirements.txt
Flask==2.3.3
firebase-admin==6.2.0
PyPDF2==3.0.1
python-docx==0.8.11
pdfplumber==0.9.0
gunicorn==21.2.0
python-dotenv==1.0.0
flask-cors==4.0.0
werkzeug==2.3.7